Make major bottom layer (see description)
Since OtherMill just mills the board but doesn't pour copper over the
holes. The top and the bottom layer of a vias are disconnected. In order
to reduce the hard word in soldering, I brough all traces to the bottom
layer & set copper pour for both sides of the board.